doc/requirements.txt
svgwrite
pillow
-e git+https://github.com/CatherineH/images2gif-Pillow.git#egg=images2gif-Pillow